add tezi manifest
This commit moves the manifest for building tezi from internal
toradex-tezi-platform to toradex-manifest, so we can make use of the pinned
BSP hashes.

This commit updates the hashes from toradex-tezi-platform as follows:

meta-toradex-bsp-common
	9f07b468d6cecb6fcda469866c305d37fb20f230 ->
	276a8e076225332b336e41eb2c62375ddc6989bf
meta-toradex-nxp
	91233079b4167fe0be7e1f37af67c261ef9afb78 ->
	4a074b05bad21ac438a398bb036ec0e8da41513c
meta-toradex-tegra
	f58026172d3c32a9c8026287370ae813d9f14fc5 ->
	d4ef0a4584e04b14a6ba368b8cdf8592af43d3bb
meta-toradex-tezi
	6a58e5cd2d5a84357686fe7c6b08e4033ca065dd ->
	e1d825219d627b5ded55e347ca4e9d4c78e41673

hashes of integration.xml and next.xml are the same if any are present there.

Related-to: ELB-4127
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>